Implementing isFull method
Lots of changes here:
- Creating the method convertToGrid to be able to obtain the position of
a row/col element in the grid object
- Adding the concept of TOP and BOTTOM element. We'll use them to know
if any point of the top row is connected to any point of the bottom row
- Creating the connect method, to connect two points (each represented
by a row+col combination). Here is where we set the logic to avoid
connecting to one point if it's closed or out of bounds.
- Creating the method connectFirstRowToTop, to make sure that all the
elements in the first row are connected to the top of the grid
- Creating the method connectToNeighbors, to simplify the connection of
one element to its neighbors when we open it
- Refactoring the open method: we have divided the logic into two helper
methods, one to handle the opening of the site (i.e. putting the tile to
true + increasing the count) and another to connecting to the
neighboring elements
- Finally, implementing isFull: we consider that an element is full if
it is connected to the top element