Running Percolation experiments + creating client
Modifying PercolationStats' constructor to make it run the specified
number of experiments and store its results

Also, creating a client in the main() method that follows the assignment
instructions

NOTE: We've set most of the fields of the class as private. This is
required by the assignment autograder, or otherwise it determines that
the API doesn't match the specified one (because it has extra fields)
and doesn't even grade the assignment. However, this will cause the
tests to fail, as they were setting the experiment results manually