Avoid specifying unchecked exceptions
The goal of the keyword throws is not to indicate that our system throws
that exception, but to specify that the exception:
- Does not depend on ourselves
- We aren't catching or trying to recover from it

In conclusion, it seems that the "philosophy" of throws is to make the
client aware of unexpected runtime exceptions in case he wants to
protect from them. In this case, the only exceptions that we throw are
the ones that are actively documented, and therefore the "throws" clause
is redundant. The way of notifying the client of the exception is via
the documentation