Preventing Backwash
Backwash is a problem caused by the fact that we have virtual sites at the bottom and the top, and because all the sites at the bottom row are connected to the virtual bottom.
When the system percolates, the virtual bottom will be connected to the virtual top.
As a consequence, all the sites in the bottom row will also be connected to the virtual top through this bottom, and will be marked as Full even if they aren't actually connected to the top
To prevent this, we create a second grid data-structure without virtual bottom. We'll use this second grid only to check if the elements are full; as it doesn't have bottom, a site will only be full if there is a real path to the top.