Week 3 - Checkpoint
Using this commit to mark that here is where we started with the second
phase of Week 3. The existing code obtains a score of 78/100 in the
assignment.

Point and BruteCollinear points obtain a perfect score, so we need to
focus our efforts in FastCollinear points. Anyway, at least
BruteCollinear points can underdo a big refactor, so we'll do some
modifications there too.