Extracting board copying logic to methods
Creating two auxiliar methods, copyBoardContent and copyFromOriginalAndSwapTiles.

copyBoardContent returns an int[][] that is a deep copy of the contents
of the board passed as an argument. Therefore, if the original board is
changed we won't be affected.

copyFromOriginalAndSwapTiles uses the previous method to create a copy
of the contents of the original board and swaps the row and cols
specified in the arguments. It's very useful if we want to create two
very similar boards (e.g twins or neighbors)