Completing Solver.java
Completed version of Solver.java. The trickiest part is the fact that we
use two different priority queues in the resolution part: one with the
original board and another one with its twin. This happens because only
one of them can be solved (i.e. if the twin has a solution, the original
is unsolvable), so we try to solve both at the same time and if we solve
the twin first, it means that the original board can be solved, so we
stop the process

Another important part is the fact that in each step we don't
insert in the PQ the neighbor that is identical to the node we're coming
from. By definition, that node won't bring us to the solution (on each
step we are closer to the solution than before), so by excluding it from
the PQ we save a lot of space.