Caching Manhattan distance in BoardNode
Each node is compared against several others, so being able to doing so fast (i.e. making the compareTo method execute fast) results in huge timing improvement in the algorithms. We will be caching the Manhattan distance of the main board, so we only need to compute it once. This results in a great improvement because the computation of the Manhattan distance is O(N^2)